+++ detekt-rules-complexity/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/complexity/ThresholdedCodeSmell.kt
@@ -1,4 +1,8 @@
-package io.gitlab.arturbosch.detekt.api
+package io.gitlab.arturbosch.detekt.rules.complexity
+
+import io.gitlab.arturbosch.detekt.api.CodeSmell
+import io.gitlab.arturbosch.detekt.api.Entity
+import io.gitlab.arturbosch.detekt.api.Issue
 
 /**
  * Represents a code smell for which a specific metric can be determined which is responsible
